+++ pkg/apis/flagger/v1beta1/zz_generated.deepcopy.go
@@ -22,6 +22,7 @@ limitations under the License.
 package v1beta1
 
 import (
+	v1alpha2 "github.com/fluxcd/flagger/pkg/apis/gatewayapi/v1alpha2"
 	v1alpha3 "github.com/fluxcd/flagger/pkg/apis/istio/v1alpha3"
 	v1 "k8s.io/api/core/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
@@ -337,6 +338,13 @@ func (in *CanaryService) DeepCopyInto(out *CanaryService) {
 		in, out := &in.Gateways, &out.Gateways
 		*out = make([]string, len(*in))
 		copy(*out, *in)
+	}
+	if in.GatewayRefs != nil {
+		in, out := &in.GatewayRefs, &out.GatewayRefs
+		*out = make([]v1alpha2.ParentReference, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
 	}
 	if in.Hosts != nil {
 		in, out := &in.Hosts, &out.Hosts