+++ test/gloo/test-canary.sh
@@ -112,14 +112,14 @@ done
 echo '✔ Canary initialization test passed'
 
 echo '>>> Triggering canary deployment'
-kubectl -n test set image deployment/podinfo podinfod=stefanprodan/podinfo:3.1.1
+kubectl -n test set image deployment/podinfo podinfod=ghcr.io/stefanprodan/podinfo:6.0.1
 
 echo '>>> Waiting for canary promotion'
 retries=50
 count=0
 ok=false
 until ${ok}; do
-    kubectl -n test describe deployment/podinfo-primary | grep '3.1.1' && ok=true || ok=false
+    kubectl -n test describe deployment/podinfo-primary | grep '6.0.1' && ok=true || ok=false
     sleep 10
     kubectl -n gloo-system logs deployment/flagger --tail 1
     count=$(($count + 1))
@@ -194,14 +194,14 @@ spec:
 EOF
 
 echo '>>> Triggering A/B testing'
-kubectl -n test set image deployment/podinfo podinfod=stefanprodan/podinfo:3.1.2
+kubectl -n test set image deployment/podinfo podinfod=ghcr.io/stefanprodan/podinfo:6.0.2
 
 echo '>>> Waiting for A/B testing promotion'
 retries=50
 count=0
 ok=false
 until ${ok}; do
-    kubectl -n test describe deployment/podinfo-primary | grep '3.1.2' && ok=true || ok=false
+    kubectl -n test describe deployment/podinfo-primary | grep '6.0.2' && ok=true || ok=false
     sleep 10
     kubectl -n gloo-system logs deployment/flagger --tail 1
     count=$(($count + 1))