+++ test/istio/test-canary.sh
@@ -121,14 +121,14 @@ fi
 echo '✔ Canary service custom metadata test passed'
 
 echo '>>> Triggering canary deployment'
-kubectl -n test set image deployment/podinfo podinfod=stefanprodan/podinfo:3.1.1
+kubectl -n test set image deployment/podinfo podinfod=ghcr.io/stefanprodan/podinfo:6.0.1
 
 echo '>>> Waiting for canary promotion'
 retries=50
 count=0
 ok=false
 until ${ok}; do
-    kubectl -n test describe deployment/podinfo-primary | grep '3.1.1' && ok=true || ok=false
+    kubectl -n test describe deployment/podinfo-primary | grep '6.0.1' && ok=true || ok=false
     sleep 10
     kubectl -n istio-system logs deployment/flagger --tail 1
     count=$(($count + 1))
@@ -219,14 +219,14 @@ spec:
 EOF
 
 echo '>>> Triggering B/G deployment'
-kubectl -n test set image deployment/podinfo podinfod=stefanprodan/podinfo:3.1.2
+kubectl -n test set image deployment/podinfo podinfod=ghcr.io/stefanprodan/podinfo:6.0.2
 
 echo '>>> Waiting for B/G promotion'
 retries=50
 count=0
 ok=false
 until ${ok}; do
-    kubectl -n test describe deployment/podinfo-primary | grep '3.1.2' && ok=true || ok=false
+    kubectl -n test describe deployment/podinfo-primary | grep '6.0.2' && ok=true || ok=false
     sleep 10
     kubectl -n istio-system logs deployment/flagger --tail 1
     count=$(($count + 1))
@@ -316,14 +316,14 @@ spec:
 EOF
 
 echo '>>> Triggering A/B testing'
-kubectl -n test set image deployment/podinfo podinfod=stefanprodan/podinfo:3.1.3
+kubectl -n test set image deployment/podinfo podinfod=ghcr.io/stefanprodan/podinfo:6.0.3
 
 echo '>>> Waiting for A/B testing promotion'
 retries=50
 count=0
 ok=false
 until ${ok}; do
-    kubectl -n test describe deployment/podinfo-primary | grep '3.1.3' && ok=true || ok=false
+    kubectl -n test describe deployment/podinfo-primary | grep '6.0.3' && ok=true || ok=false
     sleep 10
     kubectl -n istio-system logs deployment/flagger --tail 1
     count=$(($count + 1))